+++ rs/tests/consensus/utils/src/subnet_recovery.rs
@@ -1,10 +1,5 @@
+use crate::rw_message::{can_read_msg, can_store_msg, cert_state_makes_progress_with_retries};
 use crate::upgrade::assert_assigned_replica_version;
-use crate::{
-    rw_message::{
-        can_read_msg, can_store_msg, cannot_store_msg, cert_state_makes_progress_with_retries,
-    },
-    ssh_access::execute_bash_command,
-};
 use anyhow::bail;
 use candid::Principal;
 use canister_test::Canister;
@@ -17,8 +12,6 @@ use ic_consensus_threshold_sig_system_test_utils::{
 };
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_nns_constants::GOVERNANCE_CANISTER_ID;
-use ic_recovery::steps::Step;
-use ic_recovery::{get_node_metrics, NodeMetrics, Recovery};
 use ic_registry_subnet_type::SubnetType;
 use ic_system_test_driver::util::*;
 use ic_system_test_driver::{
@@ -27,108 +20,9 @@ use ic_system_test_driver::{
 };
 use ic_types::ReplicaVersion;
 use registry_canister::mutations::do_update_subnet::UpdateSubnetPayload;
-use serde::{Deserialize, Serialize};
 use slog::{info, Logger};
 use std::collections::BTreeMap;
 use std::time::Duration;
-use url::Url;
-
-/// break a subnet by breaking the replica binary on f+1 = (subnet_size - 1) / 3 +1
-/// nodes taken from the given iterator.
-pub fn break_subnet(
-    subnet: Box<dyn Iterator<Item = IcNodeSnapshot>>,
-    subnet_size: usize,
-    recovery: &Recovery,
-    logger: &Logger,
-) {
-    // Let's take f+1 nodes and break them.
-    let f = (subnet_size - 1) / 3;
-    info!(
-        logger,
-        "Breaking the subnet by breaking the replica binary on f+1={} nodes",
-        f + 1
-    );
-
-    let faulty_nodes = subnet.take(f + 1).collect::<Vec<_>>();
-    for node in faulty_nodes {
-        // simulate subnet failure by breaking the replica process, but not the orchestrator
-        recovery
-            .execute_ssh_command(
-                "admin",
-                node.get_ip_addr(),
-                "sudo mount --bind /bin/false /opt/ic/bin/replica && sudo systemctl restart ic-replica",
-            )
-            .expect("couldn't run ssh command");
-    }
-}
-
-/// Halt the subnet and wait until the given app node reports consensus 'is halted'
-pub fn halt_subnet(
-    app_node: &IcNodeSnapshot,
-    subnet_id: SubnetId,
-    recovery: &Recovery,
-    logger: &Logger,
-) {
-    #[derive(Debug, Deserialize, Serialize)]
-    struct Cursor {
-        #[serde(alias = "__CURSOR")]
-        cursor: String,
-    }
-
-    info!(logger, "Breaking the app subnet by halting it",);
-    let s = app_node.block_on_ssh_session().unwrap();
-    let message_str = execute_bash_command(
-        &s,
-        "journalctl -n1 -o json --output-fields='__CURSOR'".to_string(),
-    )
-    .expect("journal message");
-    let message: Cursor = serde_json::from_str(&message_str).expect("JSON journal message");
-    recovery
-        .halt_subnet(subnet_id, true, &[])
-        .exec()
-        .expect("Failed to halt subnet.");
-    ic_system_test_driver::retry_with_msg!(
-        "check if consensus is halted",
-        logger.clone(),
-        secs(120),
-        secs(10),
-        || {
-            let res = execute_bash_command(
-                &s,
-                format!(
-                    "journalctl --after-cursor='{}' | grep -c 'is halted'",
-                    message.cursor
-                ),
-            );
-            if res.map_or(false, |r| r.trim().parse::<i32>().unwrap() > 0) {
-                Ok(())
-            } else {
-                bail!("Did not find log entry that consensus is halted.")
-            }
-        }
-    )
-    .expect("Failed to detect broken subnet.");
-}
-
-/// A subnet is considered to be broken if it still works in read mode,
-/// but doesn't in write mode
-pub fn assert_subnet_is_broken(node_url: &Url, can_id: Principal, msg: &str, logger: &Logger) {
-    info!(logger, "Ensure the subnet works in read mode");
-    assert!(
-        can_read_msg(logger, node_url, can_id, msg),
-        "Failed to read message on node: {}",
-        node_url
-    );
-    info!(
-        logger,
-        "Ensure the subnet doesn't work in write mode anymore"
-    );
-    assert!(
-        cannot_store_msg(logger.clone(), node_url, can_id, msg),
-        "Writing messages still successful on: {}",
-        node_url
-    );
-}
 
 /// A subnet is considered to be healthy if all nodes in the given vector are healthy
 /// and running the given version, canisters can be installed and messages can be written and read.
@@ -183,76 +77,6 @@ pub fn assert_subnet_is_healthy(
         "Failed to read new message on {}",
         node.get_ip_addr()
     );
-}
-
-/// Assert that the given node has deleted its state within the next 5 minutes.
-pub fn assert_node_is_unassigned(node: &IcNodeSnapshot, logger: &Logger) {
-    info!(
-        logger,
-        "Asserting that node {} has deleted its state.",
-        node.get_ip_addr()
-    );
-    // We need to exclude the page_deltas/ directory, which is not deleted on state deletion.
-    // That is because deleting it would break SELinux assumptions.
-    let check = r#"[ "$(ls -A /var/lib/ic/data/ic_state -I page_deltas)" ] && echo "assigned" || echo "unassigned""#;
-    let s = node
-        .block_on_ssh_session()
-        .expect("Failed to establish SSH session");
-
-    ic_system_test_driver::retry_with_msg!(
-        format!("check if node {} is unassigned", node.node_id),
-        logger.clone(),
-        secs(300),
-        secs(10),
-        || match execute_bash_command(&s, check.to_string()) {
-            Ok(s) if s.trim() == "unassigned" => Ok(()),
-            Ok(s) if s.trim() == "assigned" => {
-                bail!("Node {} is still assigned.", node.get_ip_addr())
-            }
-            Ok(s) => bail!("Received unexpected output: {}", s),
-            Err(e) => bail!("Failed to read directory: {}", e),
-        }
-    )
-    .expect("Failed to detect that node has deleted its state.");
-}
-
-/// Select a node with highest finalization height in the given subnet snapshot
-pub fn select_download_node(
-    subnet: SubnetSnapshot,
-    logger: &Logger,
-) -> (IcNodeSnapshot, NodeMetrics) {
-    let node = subnet
-        .nodes()
-        .filter_map(|n| block_on(get_node_metrics(logger, &n.get_ip_addr())).map(|m| (n, m)))
-        .max_by_key(|(_, metric)| metric.finalization_height)
-        .expect("No download node found");
-    info!(
-        logger,
-        "Selected download node: ({}, {})",
-        node.0.get_ip_addr(),
-        node.1.finalization_height
-    );
-    node
-}
-
-/// Print ID and IP of all unassigned nodes and the first app subnet found.
-pub fn print_app_and_unassigned_nodes(env: &TestEnv, logger: &Logger) {
-    let topology_snapshot = env.topology_snapshot();
-
-    info!(logger, "App nodes:");
-    topology_snapshot
-        .subnets()
-        .find(|subnet| subnet.subnet_type() == SubnetType::Application)
-        .unwrap()
-        .nodes()
-        .for_each(|n| {
-            info!(logger, "A: {}, ip: {}", n.node_id, n.get_ip_addr());
-        });
-
-    info!(logger, "Unassigned nodes:");
-    topology_snapshot.unassigned_nodes().for_each(|n| {
-        info!(logger, "U: {}, ip: {}", n.node_id, n.get_ip_addr());
-    });
 }
 
 /// Enable Chain key and signing on the subnet using the given NNS node.
